+++ module-auth/src/main/java/com/mile/external/client/google/GoogleSocialStrategy.java
@@ -7,8 +7,8 @@ import com.mile.external.client.dto.UserLoginRequest;
 import com.mile.external.client.google.api.GoogleAccessTokenClient;
 import com.mile.external.client.google.api.GoogleUserClient;
 import com.mile.external.client.google.api.dto.GoogleUserInfoResponse;
-import com.mile.service.LoginStrategy;
-import com.mile.service.dto.UserInfoResponse;
+import com.mile.strategy.LoginStrategy;
+import com.mile.strategy.dto.UserInfoResponse;
 import feign.FeignException;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
