+++ module-auth/src/main/java/com/mile/external/client/kakao/KakaoSocialStrategy.java
@@ -8,8 +8,8 @@ import com.mile.external.client.kakao.api.KakaoAccessTokenClient;
 import com.mile.external.client.kakao.api.KakaoUserClient;
 import com.mile.external.client.kakao.api.dto.KakaoAccessTokenResponse;
 import com.mile.external.client.kakao.api.dto.KakaoUserResponse;
-import com.mile.service.LoginStrategy;
-import com.mile.service.dto.UserInfoResponse;
+import com.mile.strategy.LoginStrategy;
+import com.mile.strategy.dto.UserInfoResponse;
 import feign.FeignException;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
